Twitter Clone developed using php and mysql
Frontend is developed using Bootstrap. The backend was also developed with PHP. Mysql was used in the Database. The project primarily includes Register and Login operations. Afterwards, there are data addition, deletion and update operations. The added data appears on the personal page or home page. You can also exit after the transactions are completed.